Back-end/api-ai/requirements.txt
flask
flask-cors
python-dotenv
requests
numpy
faiss-cpu
sentence-transformers
pydantic
tiktoken
PyYAML
google-cloud-texttospeech>=2.16.3
